refactor: change gpt path and test code path
